Restructure some of the scanning code and read repo config
I moved the git clone operation out of the `find_strings` function,
because it had no place being in there in the first place. Now, the
CLI command will clone the remote repo, if necessary, and delete it
afterward.

I also introduced an intermediary function, "scan_repo", which will
first look for a config file in the repo, amend any config options
as necessary (for now, just include-paths and exclude-paths) and then
pass on the work to find_strings. I think this exposes a bit of a
friendlier "public" API, vs the "what does find strings actually do"
API we had.