Create an `Issue` class to represent problematic commits
This was a rather large shift in the code, and the start of a
larger cleanup discussed, in part, in #23. The actual changes
include:

* `Issue` is now a top level class used to represent issues found in
  scanning
* The `Bcolors` class is gone, replaced by the styling functionality of
  the `click` library
* `print_results` has been removed, in favor of the `__str__` method on
  the `Issue` class. This greatly improves the testability of the code
  and makes it easier to output everything at once and potentially even
  use a pager for output in the future.
* JSON output now uses the `as_dict` method on the `Issue` class, for
  a more clear representation of what exactly is output.

There is also one larger implied change here; I have not been explicitly
including Python 2.7 support. The one immediate place this is likely to
be a problem is in the use of the `enum` library. This was not added to
Python until 3.4. While there is a backport available via the `enum34`
package, I hesitate to add this. I would prefer to take this as the
beginning of 2.0 development, in which we drop support for Python 2.7
entirely.