Don't measure test coverage on commands; no good way to cover them
Because of the nature of the way that the commands are loaded and
executed, via `compile()` and `exec()`, the `coverage` module is
not able to detect when they are loaded and run.

They also can't easily be executed directly in the tests because
they require an active click context. So they are entirely executed
via the click CliRunner actually running the command.

Therefore it's not really possible to measure their coverage. So,
we'll just exclude them for now.