Split the functionality into sub-commands (#78)
* Switch to sub-command syntax for the hooks

* Extract functionality out into sub-commands

* Adjusts tests for the new GitRepoScanner signature

* Don't measure test coverage on commands; no good way to cover them

Because of the nature of the way that the commands are loaded and
executed, via `compile()` and `exec()`, the `coverage` module is
not able to detect when they are loaded and run.

They also can't easily be executed directly in the tests because
they require an active click context. So they are entirely executed
via the click CliRunner actually running the command.

Therefore it's not really possible to measure their coverage. So,
we'll just exclude them for now.

* Remove an old argument that was commented out

* Shift the tests over to the new sub-command structure

* Use click to display option defaults instead of manual

* Remove some leftover debug print statements

* Switch to using importlib instead of compile/eval

* Remove the coverage omit -- we can cover commands now!

* Change imports to make test mocking easier

* Test all the sub-commands!