Improve backward compatibility
This commit restores the `._issues` member (a list) and reverts the
`.issues` property to a list instead of a generator. Several instances
of caller references to `.issues` have been restored to their original
form.

Behind the covers, a new `._completed` member (and matching property)
tracks whether or not the scan has completed; `.issues` checks this and
forces the scan to run to completion if needed.

The low level generators now append to `._issues` instead of incrementing
`._issue_count` (which is no longer needed), and `scan()` clears state
information when it starts and then sets `._completed` when finished.

To make it all work, `echo_result()` now loops on `.scan()` instead of
`.issues` so output is still "live" although `.issues` is not.