Don't allow rescans
If `scan()` is called (again) after having completed a scan, return
immediately without doing anything. The proper way to look at issues
(again) is to consult `.issues`.

This limits, but does not eliminate, the potential for craziness if a
caller decides to interleave calls to `.scan()` and references to
`.issues`.