Feature: report issues as discovered instead of buffering (#227)
* WIP: convert issue food chain to generators

mypy is happy and unit tests have been fixed up so they pass again
(mostly by liberal application of `list()` to force execution of the
generator).

This still needs work to fix the logic where it is assumed that the
`issues` property is a list, namely where we try to take the length
of it or consume it more than once.

* Generator integration, part 2

Add new `issue_count` property to scanner class. This is incremented as
issues are yielded by the low level routines, and allows the caller to
tell how many issues we got without inspecting the pipeline multiple
times.

We needed one non-obvious fix in util.py to ensure we have consumed all
issues before checking the count. (Before, it would have been zero
nearly always.)

* Unwind default console output

Previously we marshalled all issues into a single `echo()` call, which
resulted in buffering all output until the completion of the scan.

Now we use a loop (like "compact") and issue a separate `echo()` for
each issue, allowing early output to be produced prior to the generation
of later issues. This also gets rid of the unwanted "empty output for no
issues" behavior introduced by an earlier commit.

* Stream JSON-formatted output

This works because only one part of the output is generated, so we can
output a (partial) static blob to start, then an issue at a time, and
finally terminate the object to make it valid.

Because this is a little hinky, the corresponding unit tests have been
adjusted to actually generate (and capture) the entire stdout stream.
We then deserialize (to verify syntax) and compare the resulting dict
with the original input data to confirm there was no content corruption.

* Hold issues for storage to files

Address the last remaining wart; make `echo_result()` buffer issues
(even as it streams them to console) and return the list so it's
possible to write all issues to files at the end if desired. This
returns `write_outputs()` to its original semantics, namely consumption
of a List rather than a Generator.

Arguably it might be nicer to create files as we go, but that would
constitute a behavior change (files not all done right at the end) and
need some code refactoring.

* Fix issue_count logic

Revise test to reflect use of `-1` vs `None` to indicate scan not started

* Add CHANGELOG entry

* black fixup

* rebase fixups

* Improve backward compatibility

This commit restores the `._issues` member (a list) and reverts the
`.issues` property to a list instead of a generator. Several instances
of caller references to `.issues` have been restored to their original
form.

Behind the covers, a new `._completed` member (and matching property)
tracks whether or not the scan has completed; `.issues` checks this and
forces the scan to run to completion if needed.

The low level generators now append to `._issues` instead of incrementing
`._issue_count` (which is no longer needed), and `scan()` clears state
information when it starts and then sets `._completed` when finished.

To make it all work, `echo_result()` now loops on `.scan()` instead of
`.issues` so output is still "live" although `.issues` is not.

* Don't allow rescans

If `scan()` is called (again) after having completed a scan, return
immediately without doing anything. The proper way to look at issues
(again) is to consult `.issues`.

This limits, but does not eliminate, the potential for craziness if a
caller decides to interleave calls to `.scan()` and references to
`.issues`.

* linter fixups

We squelch "too many instance attributes" on `ScannerBase` because we
really really need a `_completed` flag so we tell pylint to suck it up.

* Fix scan() re-call

Slow burn from @tarkatronic review comment. If `scan()` is called again,
yield from the cached `._issues` list instead of doing nothing. This
looks like the old behavior except it doesn't actually rescan.

* linter fixups

* Remove unnecessary mock properties

The real object doesn't have `issue_count` any longer.

* Hoist issue accumulation

Cache issues inside `scan()` instead of the lower-level routines (which
now are generators without side effects).

* Add locking for scan() method

Multiple competing calls to `scan()` are now guaranteed to execute
sequentially.

Co-authored-by: Scott Bailey <scott.bailey@godaddy.com>