Fix repository chunk generation
This addresses #243. We had two problems:

1. Diff was generated in backwards order; it is clear from:

https://pygit2.readthedocs.io/en/latest/recipes/git-show.html

that if "show diff" is implemented as:

```
>>> diff = repo.diff(commit.parents[0], commit)
```

then the previous commit must be the first argument (and in fact you can
see just above where prev_commit is precisely commit.parents[0]).

2. Commit metadata was retrieved for the previous commit

I verified running tartufo against itself that the reported commit hashes
and diffs now match what "git show" actually displays.

Unit test changes were restricted to swapping the order of the commits in
expected call signatures for repo.diff().

I resisted the temptation to collapse the special-case logic for the
lead commit into the earlier block that looks for it, in order to minimize
the size of this commit.