Fix scan-remote-repo --branch
The underlying infrastructure requires the branch to be prefixed with
the remote (i.e. "origin/"). Because we know the clone will define the
target repository as "origin", we simply modify the branch reference
when setting up parameters and everything is good.

BTW, the last time this worked correctly was v2.0.2. Tartufo versions
v2.2.0 through v2.7.0 returned no errors, but also never found any
secrets, even when they should have done so. Beginning with v2.8.0,
you get an error stating the branch cannot be found.

Fixes: #247