Fix scan-remote-repo --branch logic
A fresh-cloned repository does not have branch information for remote
branches. We must follow the information in `remotes.origin.refs`
instead of `branches` (which will have only the main/master branch).

We add a new `is_remote` flag so the common code knows which strategy
to employ.

A side-effect appears to be that because HEAD appears in the reference
list, we'll scan the default branch twice, but I can live with this.