Fix remote branch (#248)
* Fix scan-remote-repo --branch

The underlying infrastructure requires the branch to be prefixed with
the remote (i.e. "origin/"). Because we know the clone will define the
target repository as "origin", we simply modify the branch reference
when setting up parameters and everything is good.

BTW, the last time this worked correctly was v2.0.2. Tartufo versions
v2.2.0 through v2.7.0 returned no errors, but also never found any
secrets, even when they should have done so. Beginning with v2.8.0,
you get an error stating the branch cannot be found.

Fixes: #247

* Add changelog entry

* Don't break non-branch operation

Only apply magic branch name fixup if there is a name to fix up;
otherwise, pass `None` through as-is.

* Don't guess remote repository name

Thus far, when `scan-remote-repo` clones the target into a scratch
directory, the target's remote reference will be `origin`. For purposes
of future-proofing, look up and return the origin name so the caller
does not have to assume it's "origin".

* Missed adding to last commit :-p

* Unit test tweaks

Try to make path generic (like following test) to avoid spurious
failures because we can't agree on path syntax.

Co-authored-by: Scott Bailey <scott.bailey@godaddy.com>