Convert _rules_regexes from dict to set
There's really no reason for it to be a dict. This change makes it more
consistent with similar inputs and simplifies some corner cases. Note
the uniqueness test now is based on the actual pattern rather than the
name (reason), and we no longer warn about duplicates because you
can't overwrite a rule with one that does something different.