Deprecate --rules in favor of rule-patterns (#287)
* Start migrating toward rule-patterns vs rules

* Typing fixups

Adjustments to make mypy and pytest happy.

* Convert _rules_regexes from dict to set

There's really no reason for it to be a dict. This change makes it more
consistent with similar inputs and simplifies some corner cases. Note
the uniqueness test now is based on the actual pattern rather than the
name (reason), and we no longer warn about duplicates because you
can't overwrite a rule with one that does something different.

* Hide rule-patterns from the command line help

* Make configure_regexes calls more specific

* Fix bug with scanning origin/HEAD

* Test new rule-patterns config loading

* Document the new rule-patterns configuration feature

Co-authored-by: Scott Bailey <scott.bailey@godaddy.com>