Rework poetry for python 3.6
So forcing a latest-and-greatest poetry for python 3.6 doesn't work
because the latest and greatest version isn't great enough. Fall back to
letting pip install the poetry version it wants, but then additionally
install setuptools explicitly; this way, if poetry fumbles and thinks it
won't be needed, it's okay because we already installed it.