setuptools vs poetry, plan c
Forcing setuptools installation manually before running poetry is
ineffective, because the old poetry for python 3.6 removes the
"unnecessary" package.

In the key area, namely the python unit test section of tox.ini, try
telling poetry we want setuptools also to see if we can avoid this
behavior.

FWIW, I am beginning to suspect this entire situation is due to
rebuilding poetry.lock with a new poetry version and then letting an old
poetry version try to interpret it. The problem only arises with python
3.6, because other pythons support versions new enough to not be
braindamaged.