Minor Dockerfile optimizations
* Instead of having poetry tell pip how to install packages, just make
  it do the job itself.
* Since we don't consume tarball dists, don't generate them.

If you are wondering why we don't just do a single `poetry install` to
do all of the prerequisites *AND* tartufo itself, it's because it
doesn't quite work. I think there are linkages to the original content
in /app that isn't carried through to the final image; making that work
more sanely is left as an exercise for the student.