Fix libgit2 1.4.3 (intentional) regression
libgit2 version 1.4.3 introduced a number of new sanity checks aimed at
thwarting attacks where an untrusted entity could inject configuration
files into a repository that might be read and acted upon by an unwary
consumer. By default, it refuses to read a repository that is not owned
by the current user.

This change became visible in pygit2 1.9.2, which bumped its internal
libgit2 linkage from 1.4.2 to 1.4.3. Initially, 1.9.2 was kind enough to
report explicitly when raising a GitError that "repository path ... is
not owned by current user" -- but later versions did not include this
information, making it very difficult to determine why the "[r]epository
not found at ..." error was occurring.

A new option, `GIT_OPT_SET_OWNER_VALIDATION`, was added to libgit2 to
allow users to opt out of this behavior, but it was not included in
pygit2 1.9.2; luckily by the time we get to pygit2 1.11.x, it is
visible.

This commit explicitly forces `GIT_OPT_SET_OWNER_VALIDATION` to `0`,
restoring compatibility with pygit2 1.9.1 and earlier.

While there are many cases where the default behavior might be
appropriate, as a practical manner we must opt out to support use from a
docker container, where the owner of the host files will almost never be
the same as the process running tartufo inside the container. It's much
more straightforward to preserve existing behavior in all cases.

If you are a paranoid, do not run scans of repositories you do not own.
Note that remote repositories were not affected by this change, because
they always caused a clone operation where the resulting working
directory is created (and therefore owned) by the current user.