Add support for python 3.11 (#403)
* Update actions to latest versions

* actions/checkout bumped to v3.1.0
* actions/cache bumped to v3.0.11
* actions/setup-python bumped to v4.3.0

* Bump setup-buildx-action

docker/setup-buildx-action now v2.2.1 (was v1.6.0)

* Docker build fixes

* Adjust pygit2 requirements to match latest support matrix
* Bump poetry to latest stable version

* Update CHANGELOG.md

* Add formal python 3.11 support

* Add 3.11 to various test matrix
* Bump Dockerfile base from 3.10 to 3.11

* Dependency updates

* Drop references to python 3.6 in test matrices
* Incorporate all of the bot PR version bumps
* Skip GitPython over all of the yanked versions

* pylint cleanups

* Bump pylint to latest version -- eliminates numerous false positive
  detections when run with python 3.11
* Bump minimum python to 3.7.2 to placate pylint after update
* Remove exclusions for deprecated/deleted pylint checks
* Minor reworking of some tests to eliminate new pylint complaints

* Update deprecated signatures

* Bump pygit2

We need the latest version for python 3.11 support, so bump version but
pin python 3.7 to previous version, which is the last one that supports
the old python.

* Add docker build sanity check

We've had several instances where the docker build is "successful" but
the resulting container image is completely brain-damaged. Add a step
that exercises it by running a scan of the working directory used to
build the container.

* Minor Dockerfile optimizations

* Instead of having poetry tell pip how to install packages, just make
  it do the job itself.
* Since we don't consume tarball dists, don't generate them.

If you are wondering why we don't just do a single `poetry install` to
do all of the prerequisites *AND* tartufo itself, it's because it
doesn't quite work. I think there are linkages to the original content
in /app that isn't carried through to the final image; making that work
more sanely is left as an exercise for the student.

* Bump tartufo-action

Latest tagged version understands our updated signatures

* Fix docker verification step

Consolidate steps so we actually have the image available for testing

* Docker futzing

* Docker build RTFM

Hopefully this will save the image so we can test it

* Update CHANGELOG

* Fix libgit2 1.4.3 (intentional) regression

libgit2 version 1.4.3 introduced a number of new sanity checks aimed at
thwarting attacks where an untrusted entity could inject configuration
files into a repository that might be read and acted upon by an unwary
consumer. By default, it refuses to read a repository that is not owned
by the current user.

This change became visible in pygit2 1.9.2, which bumped its internal
libgit2 linkage from 1.4.2 to 1.4.3. Initially, 1.9.2 was kind enough to
report explicitly when raising a GitError that "repository path ... is
not owned by current user" -- but later versions did not include this
information, making it very difficult to determine why the "[r]epository
not found at ..." error was occurring.

A new option, `GIT_OPT_SET_OWNER_VALIDATION`, was added to libgit2 to
allow users to opt out of this behavior, but it was not included in
pygit2 1.9.2; luckily by the time we get to pygit2 1.11.x, it is
visible.

This commit explicitly forces `GIT_OPT_SET_OWNER_VALIDATION` to `0`,
restoring compatibility with pygit2 1.9.1 and earlier.

While there are many cases where the default behavior might be
appropriate, as a practical manner we must opt out to support use from a
docker container, where the owner of the host files will almost never be
the same as the process running tartufo inside the container. It's much
more straightforward to preserve existing behavior in all cases.

If you are a paranoid, do not run scans of repositories you do not own.
Note that remote repositories were not affected by this change, because
they always caused a clone operation where the resulting working
directory is created (and therefore owned) by the current user.

* Bump action versions

Update github actions to latest released versions, to see if this fixes
inscrutable linter workflow failures

* Fix tox.ini syntax

`whitelist_externals` -> `allowlist_externals`

Co-authored-by: Scott Bailey <scott.bailey@godaddy.com>