Documentation Update (#449)
* Update docs to remove references to deprecated config items


* (Maybe) fix docs generation

This is really messed up, but seems to work reliably now with local
invocations of `poetry run tox -e docs` where before it mostly failed.
As usual, I don't understand the situations under where it was able to
work before. We'll see if the CI tests actually succeed once I push
this...

* `Makefile` fixes

Extensive debugging suggested two related failure modes, where the
`sphinx-build` script could not be located or when executed it failed
to load modules. Investigation revealed that the "here" lines at the
top of `sphinx-build` and friends used relative paths
(`#!.tox/docs/bin/python`), and it appeared this was invoked in contexts
where this path was not resolvable. Tracing the actual chain of control
from `poetry` to `tox` back to `poetry` to `make` makes it difficult to
understand what is actually going on.

In any event, the fix incorporated here works around the problem by
locating the absolute location of `sphinx-build` and then invoking it
using `python /real/location/of/sphinx-build`, which causes the
problematic relative reference to be ignored because the python
interpreter is located using the scoping rules provided by `poetry`. I
believe this was the critical fix of everything incorporated here.

* `pyproject.toml` editorial fixes

These are nits, cleaning up spacing and unnecessary quoting, and have no
functional impact.

* `tox.ini` fixes

This probably isn't important, but I noted `make` wasn't getting invoked
via `poetry run` like everything else, so I fixed it.

* Fix linter error

Use explicit dict instead of casting literal input

* Revert "Remove cache. Maybe it is causing version conflict."

This reverts commit 740efe5b8022d69147d04f3b23de4d462f915299.

The cache does not appear to be at fault; let's turn it back on.

---------

Co-authored-by: Scott Bailey <scott.bailey@godaddy.com>