mypy linting fixups
The Good: use correct annotation for `value` parameter to indicate it is
a tuple consisting of an indeterminate number of strings, instead of a
tuple consisting of exactly one string.

The Bad: fix some unit tests where I forgot to convert the `value`
parameter into a tuple in the first place.

The Ugly: `read_pyproject_toml()` has been systematically abusing
click's context `default_map` by treating it like a dict even though it
actually is defined as an (immutable) `Mapping`. It's been doing this
for a long time, and evidently it works. A proper fix (probably using
`params` instead of `default_map`) is not obvious to me -- I tried just
switching to params, and it ate my updates. Therefore, we continue the
existing abusive practice, document the problem, and add an annotation
to silence the (valid) mypy error.