Reduce configuration duplicity
Why is every class doing almost identically the same thing to load
configuration? Bleah.

Add a `load_config()` method to the base `ScannerBase` class; this does
the configuration lookup once, in one place, one way.

For all of the derived classes, remove the open-coded crap (generally in
`load_repo()` even though it has nothing to do with loading the
repository) and instead make a call to `self.load_config()` from the end
of the `__init__()` method. Note that we can't simply delegate this to
the base class's initializer, because it needs to happen after the files
exist, which is after `load_repo()` in the case of remote repositories.