Fix config "overlap" brain-damage
It's not clear this really is a problem, but we had logic that
preventing reading a "local" config file if we'd already processed it
globally. However that logic was broken, and it got even more broken
when we starting allowing multiple global configuration files.

First, get rid of return value from `read_pyproject_toml()` -- which was
unrecoverable -- in favor of stashing a set of processed configuration
files globally. Make these fully-resolved Path objects to avoid any
ambiguity in what was or wasn't referenced.

While we're at it, junk the return value which provided nothing useful.
Also while we're at it, use this ourselves and avoid reusing a file if
the user says something stupid like `--config foo --config foo`.

I almost added a test to avoid re-reading a file in the case where the
user says `--config foo --config foo`, but I didn't do that so that the
scenario `--config a --config b --config a` preserves "last reference
wins" semantics, which wouldn't be the case if we short-circuited the
second read of a because we'd seen it already.

Second, replace the test against the global options -- which had the
wrong type, and which had what the user typed instead of the resolved
name, and test for membership in the processed set instead.